+++ main.go
@@ -5,8 +5,8 @@ package main
 
 import (
 	log "github.com/hashicorp/go-hclog"
-	"github.com/hashicorp/nomad-driver-ecs/ecs"
 	"github.com/hashicorp/nomad/plugins"
+	"github.com/ryadavDeqode/nomad-driver-ray/ray"
 )
 
 func main() {
@@ -16,5 +16,5 @@ func main() {
 
 // factory returns a new instance of a nomad driver plugin
 func factory(log log.Logger) interface{} {
-	return ecs.NewPlugin(log)
+	return ray.NewPlugin(log)
 }