+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 BUG FIXES:
 
-* config: Create ECS task with the the value for `assign_public_ip` as specified in the job [[GH-11](https://github.com/hashicorp/nomad-driver-ecs/pull/11)]
+* config: Create ECS task with the the value for `assign_public_ip` as specified in the job [[GH-11](https://github.com/ryadavDeqode/nomad-driver-ray/pull/11)]
 
 ## 0.1.0 (May 12, 2021)